# Performance improvements for legacy quants on ARM_NEON